リファクタリング: バックグラウンドスクリプトをモジュールコンポーネントに再構築
